Add a Score value to Statistics
The Score is a representative value that determines the professional
degree of the user. The higher the Score the better the user is in typing.
It gets higher if the user types faster and lower for every error that is
produced. The score is also proportional to the content length of the
Lesson.